Version 3.1.0 - Update UsersControllers -> Deposite function

+++ Domain/Transaction.cs
@@ -32,7 +32,7 @@
 
 		public decimal Amount { get; private set; }
 
-		public int TimeDurationInMillisecond { get; set; }
+		public double TimeDurationInMillisecond { get; set; }
